TIP-842: Move attribute repository to Akeneo\Pim\Structure (enrich bundle)
